sysarray.txt FULL version instead of mini-stub
sysarray.txt used to be just containing "0 0 8192" and then being autonomously generated to a large "full size".

Unfortunately, newer gfortran fails with this approach. It does not "branch" when the file is too short, but "fails" with a past end of record reading. For that reason, I am now uploading an ENTIRE sample sysarray.txt file, which is having its full bombastic size of 3.5 MB. (On Linux, still nicely readable with Leafpad.)